Renames lifebit to life_device
Renames the lifebit class and related files to life_device for clarity and consistency.

Updates include renaming the source and header files, adjusting include statements, and updating class references.
