+++ PAC/common/device/life_device.h
