+++ test/device/test_life_device.cpp
@@ -1,0 +1,1 @@
+#include "test_life_device.h"
