--- test/device/test_lifebit.cpp
@@ -1,1 +1,0 @@
-#include "test_lifebit.h"
